feat(services): move ArcGIS sources out of WMS
