+++ docs/developer-guide/using-sources.mdx
@@ -32,7 +32,7 @@ This means that applications written against that interface can now support the 
 
 | Source Interface   | Data access model                         | Examples                                                    |
 | ------------------ | ----------------------------------------- | ----------------------------------------------------------- |
-| `ImageSource`      | Loads image covering a region             | `WMSSourceLoader`, `_ArcGISImageServerSourceLoader`         |
+| `ImageSource`      | Loads image covering a region             | `WMSSourceLoader`, `ArcGISImageServerSourceLoader`         |
 | `RasterSource`     | Loads typed raster samples for a viewport | `GeoTIFFSourceLoader`, `OMETiffSourceLoader`                |
 | `VectorSource`     | Load "features" in a region               | WFS (N/A), ArcGIS FeatureServer                             |
 | `ImageTileSource`  | Load image covering a specific tile index | WMTS (N/A)                                                  |
