+++ docs/modules/services/README.md
@@ -1,0 +1,12 @@
+# ArcGIS Services
+
+The `@loaders.gl/services` module provides framework-independent sources for ArcGIS REST services.
+
+It supports FeatureServer, ImageServer, MapServer, and VectorTileServer endpoints. OGC protocols
+such as WMS, WMTS, WFS, GML, and CSW remain in [`@loaders.gl/wms`](/docs/modules/wms).
+
+## Installation
+
+```bash
+npm install @loaders.gl/services @loaders.gl/core
+```
