+++ docs/modules/wms/README.md
@@ -20,7 +20,6 @@ The `@loaders.gl/wms` module provides support for a subset of the OGC Web Servic
 | [**WMS**](/docs/modules/wms/formats/wms) (Web Map Service) protocol                 | Y            | protocol for serving geo-referenced map images over the internet.                                                                    |
 | [**WFS**](/docs/modules/wms/formats/wfs) (Web Feature Service) protocol             | experimental | protocol for serving geo-referenced map features (geometries) over the internet.                                                     |
 | [**WMTS**](/docs/modules/wms/formats/wmts) (Web Map Tile Service) protocol          | experimental | protocol for serving pre-rendered or run-time computed georeferenced map tiles over the Internet.                                    |
-| **ArcGIS MapServer**                                                                  | experimental | protocol for serving cached map tiles through the ArcGIS REST API.                                                                  |
 | [**GML**](/docs/modules/wms/formats/gml) (Geographic Markup Language) format        | experimental | an XML grammar that describes geographical features.                                                                                 |
 | [**WCS**](/docs/modules/wms/formats/wcs) (Web Coverage Service)                     | N            | Load coverage data (e.g. geotiff images for satellite data) from a server.                                                           |
 | [**WMC**](/docs/modules/wms/formats/wmc) (Web Map Context)                          | No           | Used in WMS clients to save the configuration of maps and to load them again later. Can also be exchanged between different clients. |
