+++ modules/services/README.md
@@ -2,8 +2,8 @@
 
 Framework-independent sources for geospatial service APIs.
 
-This package is the home for service integrations, starting with ArcGIS REST services including FeatureServer, ImageServer, MapServer, and VectorTileServer sources. OGC protocol implementations remain in `@loaders.gl/wms`; STAC and other catalog services are intentionally out of scope for this initial package.
+This package is the home for ArcGIS REST service integrations including FeatureServer, ImageServer, MapServer, and VectorTileServer sources. OGC protocol implementations remain in `@loaders.gl/wms`; STAC and other catalog services are intentionally out of scope for this initial package.
 
-The initial release provides a focused package boundary while the implementations remain shared with `@loaders.gl/wms`; existing WMS imports continue to work during the migration.
+ArcGIS sources are owned and exported exclusively by this package. OGC protocol implementations, including WMS, WMTS, WFS, GML, and CSW, are exported by `@loaders.gl/wms`.
 
 For documentation, visit the [loaders.gl website](https://loaders.gl/docs).
