+++ modules/services/src/arcgis/arcgis-image-server-source-loader.ts
