+++ modules/services/src/arcgis/arcgis-image-tile-source-loader.ts
