+++ modules/services/src/arcgis/arcgis-map-tile-source-loader.ts
