+++ modules/services/tsconfig.json
@@ -7,5 +7,11 @@
     "rootDir": "src",
     "outDir": "dist"
   },
-  "references": [{"path": "../wms"}]
+  "references": [
+    {"path": "../gis"},
+    {"path": "../images"},
+    {"path": "../lerc"},
+    {"path": "../loader-utils"},
+    {"path": "../schema"}
+  ]
 }
