+++ modules/wms/src/index.ts
@@ -118,48 +118,6 @@ export {DEFAULT_SERVICE_LOADERS, ServiceRequestError, ServiceRuntime} from './se
 export type {ServiceEndpoint, ServiceEndpointPreferences} from './capability-graph';
 export {CapabilityGraph, discoverServiceGraph} from './capability-graph';
 
-// ArcGIS SourceLoaders
-
-export {getArcGISServices as _getArcGISServices} from './arcgis/arcgis-server';
-export {ArcGISFeatureServerSourceLoader as _ArcGISFeatureServerSourceLoader} from './arcgis/arcgis-feature-server-source-loader';
-export {ArcGISImageServerSourceLoader as _ArcGISImageServerSourceLoader} from './arcgis/arcgis-image-server-source-loader';
-export type {
-  ArcGISVectorTileServiceMetadata,
-  ArcGISVectorTileServerSourceLoaderOptions
-} from './arcgis/arcgis-vector-tile-server-source-loader';
-export {
-  ArcGISVectorTileServerSource,
-  ArcGISVectorTileServerSourceLoader
-} from './arcgis/arcgis-vector-tile-server-source-loader';
-export {
-  ArcGISFeatureServerSourceLoader,
-  ArcGISVectorSource
-} from './arcgis/arcgis-feature-server-source-loader';
-export type {
-  ArcGISFeatureServiceQueryOptions,
-  ArcGISFeatureServerSourceLoaderOptions
-} from './arcgis/arcgis-feature-server-source-loader';
-export {
-  ArcGISImageServerSourceLoader,
-  ArcGISImageSource
-} from './arcgis/arcgis-image-server-source-loader';
-export type {
-  ArcGISExportImageParameters,
-  ArcGISImageSourceLoaderProps
-} from './arcgis/arcgis-image-server-source-loader';
-export {getArcGISServices} from './arcgis/arcgis-server';
-export type {Service} from './arcgis/arcgis-server';
-export type {ArcGISMapTileSourceLoaderOptions} from './arcgis/arcgis-map-tile-source-loader';
-export {
-  ArcGISMapTileSourceLoader,
-  ArcGISMapTileSource
-} from './arcgis/arcgis-map-tile-source-loader';
-export type {ArcGISImageTileSourceLoaderOptions} from './arcgis/arcgis-image-tile-source-loader';
-export {
-  ArcGISImageTileSourceLoader,
-  ArcGISImageTileSource
-} from './arcgis/arcgis-image-tile-source-loader';
-
 export {ImageSource} from '@loaders.gl/loader-utils';
 export type {ImageType} from '@loaders.gl/images';
 export {createImageSource} from './lib/deprecated/create-image-source';
