+++ modules/wms/src/service-runtime.ts
@@ -4,11 +4,6 @@
 
 import type {CoreAPI, DataSource, DataSourceOptions, SourceLoader} from '@loaders.gl/loader-utils';
 import {CSWSourceLoader} from './csw-source-loader';
-import {ArcGISFeatureServerSourceLoader} from './arcgis/arcgis-feature-server-source-loader';
-import {ArcGISImageServerSourceLoader} from './arcgis/arcgis-image-server-source-loader';
-import {ArcGISMapTileSourceLoader} from './arcgis/arcgis-map-tile-source-loader';
-import {ArcGISImageTileSourceLoader} from './arcgis/arcgis-image-tile-source-loader';
-import {ArcGISVectorTileServerSourceLoader} from './arcgis/arcgis-vector-tile-server-source-loader';
 import {WMSSourceLoader} from './wms-source-loader';
 import {WMTSSourceLoader} from './wmts-source-loader';
 import {WFSSourceLoader} from './wfs-source-loader';
@@ -70,13 +65,8 @@ export class ServiceRequestError extends Error {
   }
 }
 
-/** Default source set spanning OGC and ArcGIS service families. */
+/** Default source set spanning the OGC service families implemented by WMS. */
 export const DEFAULT_SERVICE_LOADERS: readonly ServiceSourceLoader[] = [
-  ArcGISFeatureServerSourceLoader,
-  ArcGISImageServerSourceLoader,
-  ArcGISImageTileSourceLoader,
-  ArcGISVectorTileServerSourceLoader,
-  ArcGISMapTileSourceLoader,
   WMTSSourceLoader,
   WMSSourceLoader,
   WFSSourceLoader,
