+++ modules/wms/test/index.ts
@@ -31,7 +31,5 @@ import './gml/gml-loader.spec';
 // Services
 
 import './wms/wms-source.spec';
-import './arcgis/arcgis-server.spec';
-import './arcgis/arcgis-vector-tile-server-source.spec';
 import './service-capabilities.spec';
 import './ogc-api-source.spec';
