+++ website/package.json
@@ -54,6 +54,7 @@
     "@loaders.gl/scene": "workspace:*",
     "@loaders.gl/schema": "workspace:*",
     "@loaders.gl/schema-utils": "workspace:*",
+    "@loaders.gl/services": "workspace:*",
     "@loaders.gl/splats": "workspace:*",
     "@loaders.gl/traces": "workspace:*",
     "@loaders.gl/xml": "workspace:*",
